Move ProfileCard component to Profile folder
